+++ tests/unittests/report/report_tests.py
@@ -110,18 +110,18 @@ class QualityReportTest(unittest.TestCase):
         self.assertEqual([product], quality_report.products())
 
     def test_report_with_all_metrics_green(self):
-        """ Test that the report passes the quality gate. """
+        """ Test that the report needs to direct action when all metrics are green. """
         quality_report = report.QualityReport(self.__project)
         quality_report.sections()  # Generate the report
-        self.assertTrue(quality_report.all_green())
+        self.assertFalse(quality_report.direct_action_needed())
 
-    def test_report_with_missing_metrics_has_not_all_green(self):
+    def test_direct_action_is_needed_when_metrics_are_missing(self):
         """ Test a report with a missing metric. """
         project = domain.Project('organization', name='project title',
                                  added_requirements={requirement.TrackSonarVersion, requirement.Java})
         quality_report = report.QualityReport(project)
         quality_report.sections()  # Generate the report
-        self.assertFalse(quality_report.all_green())
+        self.assertTrue(quality_report.direct_action_needed())
 
 
 class QualityReportMetaDataTest(unittest.TestCase):