+++ docs/CHANGES.md
@@ -1,3 +1,8 @@
+2017-02-12  Release 1.77.2
+
+  * Return exit status 2 if any metric is red or missing and the --failure-exit-code command line option is passed.
+  
+  
 2017-02-10  Release 1.77.1
 
   * Color the report date yellow or red, depending on the age of the report.